Working NFS over Ethernet on IGEP
The supplied buildroot config supports the use of NFS over the
Ethernet connection.

To use, un-tar your rootfs in SCRATCH on kenny and replace the
igep.ini on your boot SD-Card with the one in the configs/
subdirectory, and modify to point to your rootfs. As long as the
kernel is configured to allow rootfs to be mounted over NFS, this
should work as is.

README.nfs explains how this is achieved.