Fix coding standard
Improve code readability by aligning code blocks to 8 space tabbing.
Open braces for code blocks should be on the same line as the for/if
or other statement that requires the code block, however functions should
have an open brace on a new line for dealing with long parameter lists.