Refactor LCDDisplay output
LCDDisplay::refresh() has a huge amount of code that is duplicated
across multiple string outputs.

This code has been refactored and generalised into ::writeBlock()
which should be a drop in replacement for all of the tokenising and
outputting code found in ::refresh().

Display blocks are represented by the start location and length available
to display within the block. struct displayBlock contains this information.