Encapsulate data within Display to prevent multithreading corrupting data
The IO and main threads both access the Display object, which in turn
either use or modify the strings associated with display blocks. If
the main thread altered a string while the IO thread was displaying it,
the pointer being used by the IO thread would become invalid because it
will be freed by the main thread.

To prevent this from happening, the strings are now private and must be accessed
via getters. These getters use mutexes to prevent access from multiple locations
at once, and provide a copy of the string instead of the actual string.

Changes had to be made to subclasses to support the new interface - ::refresh()
must now use getters and setters to modify the parent classes data.