Response data is now processed into appropriate strings and vectors.
The login() function now handles error codes from the CDD, aswell as
sorts the token into a string.

If an error occurs, the login function now returns a negative code.
Calling getErr() will return a description of the error.

The knowledgelevel ID's and Descriptions are stored in two separate vectors.
The ID's and Descriptions share the same locations in each vector.
This is so that the descriptions can be fetched without containing an ID code,
and ID's can be used for creating requests to the CDD.

The same applies to the language ID's and descriptions.

The test application has been updated to reflect this, and will now print
either a token and list of supported languages / knowledgelevels or
an error message.